requirements.txt
flask==2.3.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.0.3
reportlab==4.0.4
mysql-connector-python==8.2.0
